+++ wallets/react-wallet-v2/src/lib/SmartAccountLib.ts
@@ -3,13 +3,33 @@ import { privateKeyToSafeSmartAccount } from 'permissionless/accounts'
 import { sepolia } from 'viem/chains'
 import { privateKeyToAccount } from 'viem/accounts'
 import { createWalletClient } from 'viem'
-import {
-  publicClient,
-  bundlerClient,
-  pimlicoBundlerTransport,
-  walletConnectTransport
-} from './PimlicoLib'
+import { createPimlicoBundlerClient } from 'permissionless/clients/pimlico'
+import { createPublicClient, http } from 'viem'
 
+// -- Helpers --------------------------------------------------------------------------------------
+const projectId = process.env.NEXT_PUBLIC_PROJECT_ID
+const pimlicoKey = process.env.NEXT_PUBLIC_PIMLICO_KEY
+
+// -- RPC CLient -----------------------------------------------------------------------------------
+export const pimlicoBundlerTransport = http(
+  `https://api.pimlico.io/v1/sepolia/rpc?apikey=${pimlicoKey}`
+)
+
+export const walletConnectTransport = http(
+  `https://rpc.walletconnect.com/v1/?chainId=EIP155:11155111&projectId=${projectId}`
+)
+
+export const publicClient = createPublicClient({
+  chain: sepolia,
+  transport: walletConnectTransport
+})
+
+export const bundlerClient = createPimlicoBundlerClient({
+  chain: sepolia,
+  transport: pimlicoBundlerTransport
+})
+
+// -- Smart Account --------------------------------------------------------------------------------
 export async function getOrCreateSmartAccount(signerPrivateKey: `0x${string}`) {
   // Step 1: Create smart account details and client
   const smartAccount = await privateKeyToSafeSmartAccount(publicClient, {
